Move generation features to Legacy-only, pending for Swagger
