+++ lib/pacto.rb
@@ -48,7 +48,6 @@ require 'pacto/investigation'
 require 'pacto/hooks/erb_hook'
 require 'pacto/observers/stenographer'
 require 'pacto/generator'
-require 'pacto/generator/filters'
 require 'pacto/contract_files'
 require 'pacto/contract_set'
 require 'pacto/uri'
