+++ lib/pacto/actors/json_generator.rb
@@ -5,7 +5,7 @@ module Pacto
       def build_request(contract, values = {})
         data = contract.request.to_hash
         data['uri'] = contract.request.uri(values)
-        data['body'] = JSON::Generator.generate(data['schema'])
+        data['body'] = JSON::Generator.generate(data['schema']) if data['schema']
         data['method'] = contract.request.http_method
         Pacto::PactoRequest.new(data)
       end
